test-requirements.txt
vcrpy-unittest
unittest2
odoo-test-helper
